Comply with several Chocolatey guidelines
This packages uses administrative functions but does not have the admin
tag as it should have, see Chocolatey package-validator's documentation:
https://github.com/chocolatey/package-validator/wiki/AdminTagShouldBeUsedWhenUsingAdminHelper

The installation PowerShell scripts uses the Write-Chocolatey* functions
but it shouldn't, also stated in the see Chocolatey package-validator's
documentation:
https://github.com/chocolatey/package-validator/wiki/ScriptsDoNotContainDeprecatedWriteChocolateyMethods

As advised by Chocolatey moderation, try/catch blocks are not useful, a
$ErrorActionPreference = 'Stop';
should suffice.

This commit adds the "admin" tag to the Chocolatey package, it also
removes the Write-Chocolatey* functions and the try/catch blocks.

Issue: #9