fix: Update documentation URL to a valid (non-404) URL
The URL in the .nuspec files for the documentation (docsUrl) was
pointing to a 404 HTTP error.

The once valid URL was not anymore, documentation moved to
https://documentation.fusioninventory.org

It causes issues from the Package Validator Moderation Service when
pushing .nupkg to the Chocolatey Community Repository.

This commits updates the URL, fixing the issue.

Issue: #GH-21